Changed blur and name of overview_tracker to tracker_screen.dart
