--- flutter-app/lib/screens/overview_screen.dart
@@ -1,38 +1,0 @@
-import 'dart:ui';
-
-import 'package:flutter/material.dart';
-
-class OverviewScreen extends StatefulWidget {
-  const OverviewScreen({super.key});
-
-  @override
-  State<StatefulWidget> createState() {
-    return _OverviewScreenState();
-  }
-}
-
-class _OverviewScreenState extends State<OverviewScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: BackdropFilter(
-          filter: ImageFilter.blur(sigmaX: 10.0, sigmaY: 10.0),
-          child: Container(
-            alignment: Alignment.center,
-            color: Colors.black.withOpacity(
-              0.2,
-            ), // Optional: for better visibility of the blur effect
-            child: const Text(
-              'Feature under development',
-              style: TextStyle(
-                color: Colors.black,
-                fontSize: 24,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
